fix: updating the filters global component
