+++ pkg/app/src/components/OrganisationCard/modules/listTab.tsx
@@ -1,13 +1,11 @@
-import React, { useState } from 'react'
+import React from 'react'
 
 import ClearIcon from '@mui/icons-material/Clear'
 import ClearAllIcon from '@mui/icons-material/ClearAll'
-import FilterListIcon from '@mui/icons-material/FilterList'
 import {
 	Button,
 	Checkbox,
 	Divider,
-	Drawer,
 	FormControl,
 	FormControlLabel,
 	FormGroup,
@@ -18,10 +16,11 @@ import {
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 
-export function FiltersTab() {
-	const [openDrawer, setOpenDrawer] = useState(false)
-	const handleDrawerNavigation = () => setOpenDrawer((prevState) => !prevState)
-	const ListTab = () => (
+interface ComponentProps {
+	handleDrawerNavigation: () => void
+}
+export function OrganizationFiltersListTab({ handleDrawerNavigation }: ComponentProps) {
+	return (
 		<Box sx={{ width: 280 }} role="presentation">
 			<Box sx={{ my: 2, px: 4, display: 'flex', alignItems: 'center', justifyContent: 'space-between' }}>
 				<Typography fontWeight="700" variant={'body1'}>
@@ -76,25 +75,6 @@ export function FiltersTab() {
 					</Button>
 				</Box>
 			</Box>
-		</Box>
-	)
-	return (
-		<Box
-			sx={{
-				display: 'flex',
-				justifyContent: 'space-around',
-				alignItems: 'center',
-			}}
-		>
-			<Drawer anchor={'right'} open={openDrawer} onClose={() => setOpenDrawer(false)}>
-				<ListTab />
-			</Drawer>
-			<Typography sx={{ fontWeight: '700' }} variant={'body2'}>
-				Filters
-			</Typography>
-			<IconButton aria-label="filters" onClick={handleDrawerNavigation} color="inherit" sx={{ ml: 1 }}>
-				<FilterListIcon fontSize={'small'} />
-			</IconButton>
 		</Box>
 	)
 }
