+++ pkg/translations/src/app/en-US.json
@@ -138,5 +138,8 @@
     },
     "error": {
     }
+  },
+  "label": {
+    "sort_by": "Sort By"
   }
 }
